refactor(cli): separate app from cli
This required some finagling, but ultimately I think it's worth it.

The idea is that the CLI is only one possible UI of the "app". So the
app creates the contexts and initializes all of its UIs, and the CLI
commands have access to the app context.

There is a logical interdependency there, but I managed to avoid it by
moving the app context to its own package. Not ideal, but it works. :)

It's unlikely that Disco will have a non-CLI UI, but this seems like a
good pattern.