feat: use nacl/box for encryption
It became obvious that Disco would need to use asymmetric encryption,
since sensitive data will need to be exchanged between the server and
clients before the TLS connection is established. E.g. the client TLS
certificate used for mTLS must be sent securely.

I considered using FiloSottile/age, and while it has a friendly Go API,
it's not really meant for storing potentially small values as Disco
needs. Part of it is that it doesn't support authenticated
encryption[1,2]. NaCl is a widely accepted standard that does, and has a
stable and battle-tested stdlib-ish Go implementation, so I chose it
instead.

This took a while to get right, particularly regarding the chunking.
Most of the NaCl examples show encrypting small amounts of data, but I
want for Disco to also be able to store large binary files. This means
that files larger than the recommended 16KB will need to be chunked, and
each chunk encrypted separately, using a different nonce.
There will be a small overhead with storing the data, but it's not very
significant.

For example, a file with a size of 17_725_927 bytes will be split into
1081 chunks of 16428 bytes, where:
----- 4b ----- | -- 24b -- | --- 16400b --- |
 Payload size  |   Nonce   | Encrypted data |
  (uint32)     |           |                |

And the last chunk will have 14867 bytes with:
----- 4b ----- | -- 24b -- | --- 14839b --- |
 Payload size  |   Nonce   | Encrypted data |
  (uint32)     |           |                |

This means that the overhead is 28 bytes for each chunk, which in this
case is 1082*28=30_296 bytes, in addition to the 16 bytes of overhead of
each NaCl boxed message, which is 1082*16=17_312 bytes.

So the total size of the encrypted data in this case would be:
17_725_927+30_296+17_312=17_773_535 bytes

For a total overhead of 47KB, which is not so bad.

You probably don't want to store gigabytes of data in Disco, but for
binary files of an "acceptable" size, it's good enough.

A big help to arrive at this structure was this reference:
https://github.com/billyboar/GCSolutions/tree/master/april15/normal/ken-pratt/go_challenge_2

This also changes the store interface to Get and Set io.Reader instead
of []byte. This allows streaming the data when setting from stdin or the
HTTP request body.

[1]: https://github.com/FiloSottile/age/issues/51

[2]: https://news.ycombinator.com/item?id=32981959