feat(crypto): use NaCl for symmetric encryption, remove AES-GCM
After reading [1,2,3] I decided to go all in with NaCl for both
symmetric and asymmetric encryption. Disco will use asymmetric
encryption for transferring the TLS certificate to establish mTLS, and
symmetric encryption for storing the data locally.

This removes the previous AES-GCM cryptopasta example as it's not a
good fit for encrypting large amounts of data, because it doesn't
support streaming (or at least, the current implementation doesn't), so
the entire file will have to be kept in memory. Unfortunately, this is
still the case, since we have to read all encrypted data in memory
before we store it in the database, but hopefully I can find a streaming
solution for storing data in SQLite as well. This removal at least
converges on a single encryption standard, so we can use NaCl for both
use cases.

[1]: https://gist.github.com/atoponce/07d8d4c833873be2f68c34f9afc5a78a

[2]: https://crypto.stackexchange.com/a/91865

[3]: https://stackoverflow.com/q/4299795