feat(cli,db): hide local user, prevent removal
There are now two different user types: local and remote, but this
distinction is not surfaced to the end user, and the `user` command only
manages remote users. The single local user is created automatically by
the `init` command, and is reserved for the user of the CLI, but in the
future we might want to expand this and allow creating multiple local
users.