fix(db): don't use upsert for user update commands
Previously, an update command would actually create the user, which is
not what we want. This now explicitly uses the UPDATE statement for
updates to enforce this. It's a bit more code, and the SQL and dynamic
queries are getting a bit cumbersome at this point... Should I just use
sqlx?