refactor: move loading of local user and encryption key to an app option
The encryption key is also needed for the main disco database, for
encrypting the X25519 private key used for invites. So instead of
reading it in a Store option, we move this to an App option. This has
the additional benefit of being able to set a specific local user, which
will be useful for tests.