feat(db): store invite X25519 private key encrypted in memory
This is a general security improvement I should make when handling
private encryption keys. Keep keys unencrypted in memory for the
absolute shortest time, and only while they're needed. This minimizes
the chances an attacker could extract the keys from memory. The fact
they're currently passed via env var / CLI kind of defeats the purpose
of this, but hopefully I can arrive at a more secure solution to that as
well that doesn't impact usability.