feat(cli): allow changing data directories
This is needed to simplify testing of multiple environments on the same
machine, which will become useful when implementing the remote
functionality.

It also makes things neater by moving data dir creation to App.Run(),
and removing the data store initialization as app options, as those are
really app requirements. In-memory stores can still be used for tests by
passing the memoryfs VFS interface to app.WithFS().

It also cleans up and fixes the logic for reading the encryption key for
specific commands, creates a cli.CLI constructor, and improves the
separation between App and CLI.

Happy with how this turned out :)