feat(web): sign the invite token with the shared key
This improves security by making replay attacks impossible. Even if an
attacker intercepts the /join request and gets access to the static
token, they don't have access to the ephemeral shared key which signs
the token, and this is different for every request. To generate the
shared key, they would also need to have the server's public X25519 key,
which is sent out-of-band and only the trusted client should have.

This change increases the risk of DoS attacks, since previously the
/join handler would exit if the token was not found, whereas now if the
token is found the server will continue by reading the entire request
body. So an attacker could make large request to exhaust server
resources, but we can protect against this by setting a max request
size.