feat(web): generate random TLS server SAN, encrypt entire join response payload
I'm not sure if using a random (but fixed for the lifetime of the Disco
stores) TLS Subject Alternative Name is a bad practice or not. Commonly,
with certs signed by a public CA, it should contain all FQDNs of the
server. But Disco uses self-signed certs and will likely never use
publically signed certs (or be exposed to the internet), and we can't
reliably get the FQDN of the server. The SAN is sent in the join
response and all clients will store it, so in practice this isn't an
issue.

The benefit is that it improves security since it's not static for all
instances of Disco, and it makes each instance unique, even if it's
running on the same server.

For now, it works well, but it may need to be changed.