feat(web): add hybrid HTTP/TLS server listener
I'm not convinced if this is a good idea or not, but it would certainly
be convenient if the same port could be used for both HTTP and TLS.
One thing that I think could be a problem is forcing specific routes to
be served over TLS only, e.g. /api/v1/store. This would be trivial with
a split approach, without involving hacks like this, but then the client
would have to be configured with 2 different ports: one for the initial
authentication, and another for actual usage.

Though this could be simplified if the server sent the TLS port in the
authentication response, in addition to the TLS client cert... Hmm.