feat(db): use app logger in migrator instead of global logger
This actually fixes a data race I noticed while working on the logging
test. For some reason, using the global logger triggers the race
detector in some low-level Go runtime string copying functions.

This is not perfect, and ideally the logger should be a field on both
db.DB and sqlite.Store, but passing the app logger to their Open()
functions was difficult in an integration test, since we initialize the
stores before the app itself is initialized, yet the logger is created
in the app WithLogger option.

So this needs a bit of rethinking, but this change fixed the data race
issue in tests, as well as made the logging test less flaky, so it's
good enough for now.