test(crypto): remove hash test
The current bcrypt 14 work difficulty is very slow on my old Core i7
machine (~30s), so running the full test suite is needlessly slow.
HashPassword is not even used by Disco, but I don't want to remove it.
The implementation is only a single line, so coverage doesn't matter in
this case. And this is not even my code, but it was copied from
cryptopasta, so I can always add it back if needed.