ci: add Earthly workflow
I really want to minimize reliance on GitHub Actions... The mess of
YAML, and inability to run workflows locally (act[1] is hacky and
broken), makes iterating and troubleshooting CI a huge pain.

Earthly[1] is not perfect, but the combination of a Makefile and
Dockerfile is a great idea. They've had some bumps along the way[3],
but the open source product is solid, and their new approach of
integrating with existing CI systems seems more future-proof. So I'm
hopeful they won't disappear anytime soon. If they do, I can always
fallback to using a fork, or a plain Dockerfile.

I won't be able to completely eliminate usage of GitHub Actions, but the
bulk of the work will be done by Earthly.

[1]: https://github.com/nektos/act

[2]: https://earthly.dev/

[3]: https://earthly.dev/blog/shutting-down-earthly-ci/